Embedded Integrator
The IntegratorAPI module contains an abstract class Integrator, through
which all the components and other modules can communicate. Its
implementation HodikIntegratorImpl lies in the GameLogic module. It
[implementation] is registered via ServiceProvider annotation, so it can
be discovered with Lookup, which is used in the abstract class
Integrator.
This allows to connect to the Integrator from the different modules,
which do not depend on the integrator implementation and are wholly
separate from each other and from the game logic. Also there is no need
to drag the Integrator instance from one constructor to another inside
the GameLogic module. All you have to do to use the integrator in your
class is to:
-import package 'org.intsys16.integrator.api.Integrator';
-create a private element and initialize it like this:
private Integrator i = Integrator.getIntegrator(); (This is the static
method of the abstract class Integrator, which returns the integrator
instance.).